Refactor blog view and dashboard for improved structure
Simplified components and removed redundant state management for the blog view. Introduced metadata generation for dynamic SEO handling and created a reusable dashboard layout component. This improves maintainability and adheres to a modular structure.
