+++ .env.example
@@ -6,6 +6,7 @@ NODE_ENV=<NODE_ENV># production | development
 SAMESITE="<SAMESITE>" # lax | strict
 NEXT_PUBLIC_TINY_API_KEY="<NEXT_PUBLIC_TINY_API_KEY>"
 NEXT_PUBLIC_UPLOADPRESET="<NEXT_PUBLIC_UPLOADPRESET>"
+NEXT_PUBLIC_SITE_URL="<NEXT_PUBLIC_SITE_URL>"
 
 CLOUDINARY_CLOUD_NAME="<CLOUDINARY_CLOUD_NAME>"
 CLOUDINARY_API_KEY="<CLOUDINARY_API_KEY>"
