+++ src/app/(dashboard)/dashboard/blog/view/[slug]/page.tsx
@@ -1,12 +1,16 @@
-'use client'
-
-import {FC, useState} from "react";
+import {FC} from "react";
 import ViewBlogComponent from "@/components/shared/viewBlog.component";
-import {Blog} from "@/type";
+import {getBlogSlugProps} from "@/type";
+import {getBlogBySlug} from "@/lib/fetchData/blog";
+import ErrorBlogComponent from "@/components/dashboard/blogs/view/errorBlog.component";
 
-const SlugPage: FC = () => {
-   const [data, setData] = useState<Blog>()
-   return <ViewBlogComponent isLogin={true} data={data} setData={setData} />
+const SlugPage: FC<getBlogSlugProps> = async ({params}) => {
+    const {slug} = await params
+    const data = await getBlogBySlug(slug)
+
+    if (data?.message !== undefined) return <ErrorBlogComponent error={data.message}/>
+
+    return <ViewBlogComponent isLogin={true} data={data}/>
 };
 
 export default SlugPage;
