+++ src/enum.ts
@@ -1,9 +1,10 @@
 export enum ENV {
     BACKEND_URL = 'BACKEND_URL',
-    BACKEND_URL_WEB='BACKEND_URL_WEB',
+    BACKEND_URL_WEB = 'BACKEND_URL_WEB',
     NEXT_PUBLIC_BACKEND_URL = "NEXT_PUBLIC_BACKEND_URL",
-    NODE_ENV= "NODE_ENV",
-    SAMESITE="SAMESITE",
-    NEXT_PUBLIC_TINY_API_KEY="NEXT_PUBLIC_TINY_API_KEY",
-    NEXT_PUBLIC_UPLOADPRESET="NEXT_PUBLIC_UPLOADPRESET"
+    NODE_ENV = "NODE_ENV",
+    SAMESITE = "SAMESITE",
+    NEXT_PUBLIC_TINY_API_KEY = "NEXT_PUBLIC_TINY_API_KEY",
+    NEXT_PUBLIC_UPLOADPRESET = "NEXT_PUBLIC_UPLOADPRESET",
+    NEXT_PUBLIC_SITE_URL = "NEXT_PUBLIC_SITE_URL"
 }
